Adds configuration via EEPROM to AnyKey
The keycode to be sent is now configurable via EEPROM data with no need
to re-flash the firmware.
The keycode is a single byte value read from EEPROM address '1'.

This commit also fixes the Makefile for AnyKey so that all targets that
use avrdude work properly (e.g. flash, read_eeprom, ...).