Adds HID-RT app
HID-RT is a minimal example for an implementation of a
command line interface over HID feature reports.
The target device will accept variable length commands
of up to 32 bytes in length and respond (when polled)
with 0 or more 32 byte answers.
The client application is implemented as a Python script
(hid-rt.py) using python-libusb1 and has been tested
on Linux. It needs to be run as root unless the host
system is configured to allow appropriate access
to USB devices for regular users.